# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: P4 Product Manager API
  version: 1.0.0
  description: |
    REST API for the P4 Product Manager e-commerce demo (Express + Sequelize).

    **Demo accounts**
    - Admin: `admin@demo.com` / `admin123`
    - User: `user@demo.com` / `user123`

    **Response envelopes**
    - Success: `{ data, message? }`
    - Paginated: `{ data: [], meta: { page, pageSize, total, totalPages }, message? }`
    - Error: `{ message, code?, extensions? }`

    **Localization**
    All routes under `/api` accept an optional `Accept-Language` header (e.g. `vi`, `en`)
    to translate `message` and error strings.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://p4-api.couponlinkh.com
    description: Production
...
security: []
paths:
...
  /api/auth/google:
    post:
      tags: [Auth]
      summary: Login with Google ID token (storefront)
      operationId: loginGoogle
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [credential]
              properties:
                credential:
                  type: string
                  description: Google Identity Services ID token
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getMe
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateMe
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/password:
    patch:
      tags: [Auth]
      summary: Set or change the current user's password
      operationId: updatePassword
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordInput'
      responses:
        '200':
          description: Password updated and a new token pair issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/products:
    get:
      tags: [Products]
      summary: List products (public, paginated)
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SearchQuery'
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by category ID
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
    post:
      tags: [Products]
      summary: Create product (admin)
      operationId: createProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Products]
      summary: Update product (admin)
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInputPartial'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Products]
      summary: Delete product (admin)
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/products/{id}/images:
    post:
      tags: [Products]
      summary: Add product gallery image (admin)
      operationId: addProductImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductImageInput'
      responses:
        '201':
          description: Image added
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductImage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/{id}/images/{imageId}:
    delete:
      tags: [Products]
      summary: Remove product gallery image (admin)
      operationId: removeProductImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/ProductImageId'
      responses:
        '200':
          description: Image removed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductImage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/refunds:
    get:
      tags: [Refunds]
      summary: List refunds (admin)
      operationId: listRefunds
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - name: orderId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by order ID
      responses:
        '200':
          description: Refund list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Refund'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/addresses:
    get:
      tags: [Addresses]
      summary: List current user's addresses
      operationId: listAddresses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Address list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Addresses]
      summary: Create address
      operationId: createAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInput'
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/addresses/{id}:
    put:
      tags: [Addresses]
      summary: Update address
      operationId: updateAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/AddressId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressInputPartial'
      responses:
        '200':
          description: Address updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Addresses]
      summary: Delete address
      operationId: deleteAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/AddressId'
      responses:
        '200':
          description: Address deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Address'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login or refresh
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        example: vi
      description: Optional locale for translated messages
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SearchQuery:
      name: q
      in: query
      schema:
        type: string
        default: ''
      description: Search keyword
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Product ID (string slug)
    ProductImageId:
      name: imageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AddressId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        message:
          type: string
    PaginatedEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        message:
          type: string
    PaginationMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ApiError:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        extensions:
          type: object
          additionalProperties: true
    Role:
      type: string
      enum: [admin, user]
    User:
      type: object
      required: [id, email, role, hasPassword]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        avatarId:
          type: string
          format: uuid
          nullable: true
        avatarUrl:
          type: string
        hasPassword:
          type: boolean
...
    LoginResponse:
      type: object
      required: [token, refreshToken, user]
      properties:
        token:
          type: string
          description: JWT access token
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UpdateProfileInput:
      type: object
      properties:
        imageStorageId:
          type: string
          description: Cloudinary public ID after client upload
    UpdatePasswordInput:
      type: object
      required: [newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          description: Required when the account already has a password
        newPassword:
          type: string
          format: password
          minLength: 6
    MeResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    Category:
      type: object
      required: [id, name, slug, description]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
...
    ProductImage:
      type: object
      required: [id, fileId, url, sortOrder]
      properties:
        id:
          type: string
          format: uuid
        fileId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        sortOrder:
          type: integer
          minimum: 0
    Product:
      type: object
      required: [id, name, price, description, imageUrl, stock]
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: integer
          minimum: 0
          description: Price in smallest currency unit
        description:
          type: string
        imageUrl:
          type: string
        imageId:
          type: string
          format: uuid
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        category:
          allOf:
            - $ref: '#/components/schemas/Category'
          nullable: true
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        stock:
          type: integer
          minimum: 0
    ProductInput:
      type: object
      required: [name, price, stock]
      properties:
        name:
          type: string
        price:
          type: integer
          minimum: 0
        description:
          type: string
        stock:
          type: integer
          minimum: 0
        imageStorageId:
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
        imageStorageIds:
          type: array
          items:
            type: string
    ProductInputPartial:
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          minimum: 0
        description:
          type: string
        stock:
          type: integer
          minimum: 0
        imageStorageId:
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
    AddProductImageInput:
      type: object
      required: [imageStorageId]
      properties:
        imageStorageId:
          type: string
        sortOrder:
          type: integer
          minimum: 0
...
    RefundStatus:
      type: string
      enum: [pending, completed, failed]
...
    Refund:
      type: object
      required: [id, paymentId, orderId, amount, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        amount:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/RefundStatus'
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Address:
      type: object
      required: [id, userId, fullName, phone, line1, city, isDefault]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        fullName:
          type: string
        phone:
          type: string
        line1:
          type: string
        city:
          type: string
        isDefault:
          type: boolean
    AddressInput:
      type: object
      required: [fullName, phone, line1, city]
      properties:
        fullName:
          type: string
        phone:
          type: string
        line1:
          type: string
        city:
          type: string
        isDefault:
          type: boolean
    AddressInputPartial:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        line1:
          type: string
        city:
          type: string
        isDefault:
          type: boolean
...